fixed 351 modal not displaying bug
